+++ src/app/view/endpoints/clusters/cluster/organization/space/detail/users/space-detail.users.module.js
@@ -40,6 +40,7 @@
     this.guid = $stateParams.guid;
     this.spaceGuid = $stateParams.space;
     this.users = [];
+    this.removingSpace = {};
     this.usersModel = modelManager.retrieve('cloud-foundry.model.users');
 
     this.organizationModel = modelManager.retrieve('cloud-foundry.model.organization');
@@ -130,7 +131,12 @@
 
     this.removeSpaceRole = function (user, spaceRole) {
       var space = that.space.details.space;
-      this.removingSpace = true;
+      var pillKey = space.entity.name + spaceRole.roleLabel;
+      if (this.removingSpace[pillKey]) {
+        return;
+      }
+      console.log('this.removingSpace[pillKey] = true; pillKey === ' + pillKey, spaceRole);
+      this.removingSpace[pillKey] = true;
       rolesService.removeSpaceRole(that.guid, space.entity.organization_guid, space.metadata.guid, user, spaceRole.role)
         .then(function () {
           return refreshUsers();
@@ -139,7 +145,7 @@
           $log.error('Failed to remove role \'' + spaceRole.roleLabel + '\' for user \'' + user.entity.username + '\'');
         })
         .finally(function () {
-          that.removingSpace = false;
+          that.removingSpace[pillKey] = false;
         });
     };
